feat(vmm): Make SerialOut into Write+Send+Debug
Remove the SerialOut specialized enum and replace it with a generic
Trait so that other outputs can be used, such as `std::fs::File`.
