+++ src/vmm/src/builder.rs
@@ -50,7 +50,7 @@ use crate::devices::virtio::net::Net;
 use crate::devices::virtio::rng::Entropy;
 use crate::devices::virtio::vsock::{Vsock, VsockUnixBackend};
 use crate::devices::BusDevice;
-use crate::logger::{debug, error};
+use crate::logger::debug;
 use crate::persist::{MicrovmState, MicrovmStateError};
 use crate::resources::VmResources;
 use crate::snapshot::Persist;
@@ -123,6 +123,7 @@ impl std::convert::From<linux_loader::cmdline::Error> for StartMicrovmError {
 }
 
 #[cfg_attr(target_arch = "aarch64", allow(unused))]
+#[allow(clippy::too_many_arguments)]
 fn create_vmm_and_vcpus(
     instance_info: &InstanceInfo,
     event_manager: &mut EventManager,
@@ -131,6 +132,7 @@ fn create_vmm_and_vcpus(
     track_dirty_pages: bool,
     vcpu_count: u8,
     kvm_capabilities: Vec<KvmCapability>,
+    serial_output: Box<dyn SerialOut>,
 ) -> Result<(Vmm, Vec<Vcpu>), StartMicrovmError> {
     use self::StartMicrovmError::*;
 
@@ -164,12 +166,9 @@ fn create_vmm_and_vcpus(
         setup_interrupt_controller(&mut vm)?;
         let vcpus = create_vcpus(&vm, vcpu_count, &vcpus_exit_evt).map_err(Internal)?;
 
-        // Make stdout non blocking.
-        set_stdout_nonblocking();
-
         // Serial device setup.
-        let serial_device =
-            setup_serial_device(event_manager, std::io::stdin(), io::stdout()).map_err(Internal)?;
+        let serial_device = setup_serial_device(event_manager, std::io::stdin(), serial_output)
+            .map_err(Internal)?;
 
         // x86_64 uses the i8042 reset event as the Vmm exit event.
         let reset_evt = vcpus_exit_evt
@@ -226,6 +225,7 @@ pub fn build_microvm_for_boot(
     vm_resources: &super::resources::VmResources,
     event_manager: &mut EventManager,
     seccomp_filters: &BpfThreadMap,
+    serial_output: Box<dyn SerialOut>,
 ) -> Result<Arc<Mutex<Vmm>>, StartMicrovmError> {
     use self::StartMicrovmError::*;
 
@@ -259,6 +259,7 @@ pub fn build_microvm_for_boot(
         track_dirty_pages,
         vm_resources.vm_config.vcpu_count,
         cpu_template.kvm_capabilities.clone(),
+        serial_output,
     )?;
 
     // The boot timer device needs to be the first device attached in order
@@ -347,9 +348,16 @@ pub fn build_and_boot_microvm(
     vm_resources: &super::resources::VmResources,
     event_manager: &mut EventManager,
     seccomp_filters: &BpfThreadMap,
+    serial_output: Box<dyn SerialOut>,
 ) -> Result<Arc<Mutex<Vmm>>, StartMicrovmError> {
     debug!("event_start: build microvm for boot");
-    let vmm = build_microvm_for_boot(instance_info, vm_resources, event_manager, seccomp_filters)?;
+    let vmm = build_microvm_for_boot(
+        instance_info,
+        vm_resources,
+        event_manager,
+        seccomp_filters,
+        serial_output,
+    )?;
     debug!("event_end: build microvm for boot");
     // The vcpus start off in the `Paused` state, let them run.
     debug!("event_start: boot microvm");
@@ -420,6 +428,7 @@ pub fn build_microvm_from_snapshot(
         vm_resources.vm_config.track_dirty_pages,
         vm_resources.vm_config.vcpu_count,
         microvm_state.vm_state.kvm_cap_modifiers.clone(),
+        Box::new(std::io::stdout()),
     )?;
 
     #[cfg(target_arch = "x86_64")]
@@ -617,7 +626,7 @@ pub fn setup_interrupt_controller(vm: &mut Vm, vcpu_count: u8) -> Result<(), Sta
 pub fn setup_serial_device(
     event_manager: &mut EventManager,
     input: std::io::Stdin,
-    out: std::io::Stdout,
+    out: Box<dyn SerialOut>,
 ) -> Result<Arc<Mutex<BusDevice>>, VmmError> {
     let interrupt_evt = EventFdTrigger::new(EventFd::new(EFD_NONBLOCK).map_err(VmmError::EventFd)?);
     let kick_stdin_read_evt =
@@ -628,7 +637,7 @@ pub fn setup_serial_device(
             SerialEventsWrapper {
                 buffer_ready_event_fd: Some(kick_stdin_read_evt),
             },
-            SerialOut::Stdout(out),
+            out,
         ),
         input: Some(input),
     })));
@@ -651,9 +660,8 @@ fn attach_legacy_devices_aarch64(
         .contains("console=");
 
     if cmdline_contains_console {
-        // Make stdout non-blocking.
-        set_stdout_nonblocking();
-        let serial = setup_serial_device(event_manager, std::io::stdin(), std::io::stdout())?;
+        let serial =
+            setup_serial_device(event_manager, std::io::stdin(), Box::new(std::io::stdout()))?;
         vmm.mmio_device_manager
             .register_mmio_serial(vmm.vm.fd(), serial, None)
             .map_err(VmmError::RegisterMMIODevice)?;
@@ -915,20 +923,6 @@ fn attach_balloon_device(
     attach_virtio_device(event_manager, vmm, id, balloon.clone(), cmdline, false)
 }
 
-// Adds `O_NONBLOCK` to the stdout flags.
-pub(crate) fn set_stdout_nonblocking() {
-    // SAFETY: Call is safe since parameters are valid.
-    let flags = unsafe { libc::fcntl(libc::STDOUT_FILENO, libc::F_GETFL, 0) };
-    if flags < 0 {
-        error!("Could not get Firecracker stdout flags.");
-    }
-    // SAFETY: Call is safe since parameters are valid.
-    let rc = unsafe { libc::fcntl(libc::STDOUT_FILENO, libc::F_SETFL, flags | libc::O_NONBLOCK) };
-    if rc < 0 {
-        error!("Could not set Firecracker stdout to non-blocking.");
-    }
-}
-
 #[cfg(test)]
 pub mod tests {
     use std::io::Write;
@@ -1034,7 +1028,7 @@ pub mod tests {
                     SerialEventsWrapper {
                         buffer_ready_event_fd: None,
                     },
-                    SerialOut::Sink(std::io::sink()),
+                    Box::new(std::io::sink()),
                 ),
                 input: None,
             }))),
